+++ dflib/src/main/java/com/nhl/dflib/BooleanSeries.java
@@ -30,6 +30,8 @@ public interface BooleanSeries extends Series<Boolean> {
 
     BooleanSeries tailBoolean(int len);
 
+    BooleanSeries selectBoolean(BooleanSeries positions);
+
     /**
      * Returns an IntSeries that represents positions in the Series that contain true values. The returned value can be
      * used to "select" data from this Series or from DataFrame containing this Series. Same as {@link #index(ValuePredicate)},