+++ dflib/src/main/java/com/nhl/dflib/DataFrame.java
@@ -203,7 +203,7 @@ public interface DataFrame extends Iterable<RowProxy> {
     DataFrame map(Index mappedColumns, RowMapper rowMapper);
 
     /**
-     * Creates a new Series with values mapped by apply row mapper function to the DataFrame. The returned Series
+     * Creates a new Series with values mapped by applying row mapper function to the DataFrame. The returned Series
      * size is the same this DataFrame height.
      *
      * @param rowMapper a function applied to each row of this DataFrame
@@ -212,6 +212,16 @@ public interface DataFrame extends Iterable<RowProxy> {
      * @since 0.6
      */
     <T> Series<T> mapColumn(RowToValueMapper<T> rowMapper);
+
+    /**
+     * Creates a new BooleanSeries with values mapped by applying row mapper function to the DataFrame. The returned
+     * series size is the same this DataFrame height.
+     *
+     * @param rowMapper a boolean function applied to each row of this DataFrame
+     * @return a new BooleanSeries.
+     * @since 0.6
+     */
+    BooleanSeries mapColumnAsBoolean(RowToBooleanValueMapper rowMapper);
 
     /**
      * Creates a new DataFrame which is the exact copy of this DataFrame, only with a single column values transformed
@@ -504,6 +514,15 @@ public interface DataFrame extends Iterable<RowProxy> {
      * @since 0.6
      */
     DataFrame selectRows(IntSeries rowPositions);
+
+    /**
+     * Returns a DataFrame with subset of rows matching condition.
+     *
+     * @param condition a {@link BooleanSeries} whose "true" values indicate which
+     * @return
+     * @since 0.6
+     */
+    DataFrame selectRows(BooleanSeries condition);
 
     /**
      * Selects DataFrame rows based on provided row index. This allows to reorder, filter, duplicate rows of this