+++ dflib/src/main/java/com/nhl/dflib/IntSeries.java
@@ -33,6 +33,8 @@ public interface IntSeries extends Series<Integer> {
 
     IntSeries tailInt(int len);
 
+    IntSeries selectInt(BooleanSeries positions);
+
     /**
      * Returns an IntSeries that represents positions in the Series that match the predicate. The returned value can be
      * used to "select" data from this Series or from DataFrame containing this Series. Same as {@link #index(ValuePredicate)},