+++ dflib/src/main/java/com/nhl/dflib/RowToBooleanValueMapper.java
@@ -1,0 +1,11 @@
+package com.nhl.dflib;
+
+import com.nhl.dflib.row.RowProxy;
+
+/**
+ * @since 0.6
+ */
+public interface RowToBooleanValueMapper {
+
+    boolean map(RowProxy row);
+}