+++ dflib/src/test/java/com/nhl/dflib/LongSeries_SelectTest.java
@@ -8,28 +8,36 @@ import static org.junit.Assert.*;
 public class LongSeries_SelectTest {
 
     @Test
-    public void test() {
+    public void testPositional() {
         Series<Long> s = LongSeries.forLongs(3, 4, 2).select(2, 1);
         new SeriesAsserts(s).expectData(2L, 4L);
         assertTrue(s instanceof LongSeries);
     }
 
     @Test
-    public void test_Empty() {
+    public void testPositional_Empty() {
         Series<Long> s = LongSeries.forLongs(3, 4, 2).select();
         new SeriesAsserts(s).expectData();
         assertTrue(s instanceof LongSeries);
     }
 
     @Test(expected = ArrayIndexOutOfBoundsException.class)
-    public void test_OutOfBounds() {
+    public void tesPositionalt_OutOfBounds() {
         LongSeries.forLongs(3, 4, 2).select(0, 3);
     }
 
     @Test
-    public void testNulls() {
+    public void testPositionalNulls() {
         Series<Long> s = LongSeries.forLongs(3, 4, 2).select(2, 1, -1);
         new SeriesAsserts(s).expectData(2L, 4L, null);
         assertFalse(s instanceof LongSeries);
+    }
+
+    @Test
+    public void testBoolean() {
+        BooleanSeries condition = BooleanSeries.forBooleans(false, true, true);
+        Series<Long> s = LongSeries.forLongs(3, 4, 2).select(condition);
+        new SeriesAsserts(s).expectData(4L, 2L);
+        assertTrue(s instanceof LongSeries);
     }
 }